Add read proxies from files

+++ proxy/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Proxy(models.Model):
     host = models.CharField(max_length=20)
-    port = models.PositiveSmallIntegerField()
+    port = models.PositiveIntegerField()
     is_active = models.BooleanField(blank=True, null=True)
     country = models.CharField(max_length=255, blank=True, null=True)
     country_code = models.CharField(max_length=5, blank=True, null=True)
